% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid1.R
\docType{data}
\name{covid1}
\alias{covid1}
\title{Comapre the everyday new cases in Australia and Canada}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 120 rows and 3 columns.
}
\source{
WHO
}
\usage{
covid1
}
\description{
This is calculated everyday new cases from 01 March 2020 to 01 May 2020 in Australia and Canada
}
\keyword{datasets}
