% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid2.R
\docType{data}
\name{covid2}
\alias{covid2}
\title{Comapre the total confirmed cases and deaths in Australia and Canada}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 558 rows and 4 columns.
}
\source{
WHO
}
\usage{
covid2
}
\description{
This is calculated total cases and deaths from 03 Jan 2020 to 07 Oct 2020 in Australia and Canada
}
\keyword{datasets}
