% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchapp.R
\docType{data}
\name{launch_app}
\alias{launch_app}
\title{Shiny for comparing COVID19 cases in Australia and Canada}
\format{
An object of class \code{function} of length 1.
}
\usage{
launch_app
}
\description{
This is visualazing and analyzing cases and deaths from 03 Jan 2020 to 07 Oct 2020 in Australia and Canada
}
\keyword{datasets}
